requirements.txt
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==22.0.0
azure-storage-blob==12.24.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.10.1
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pywin32==308
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
